requirements.txt
flask==2.3.3
google-generativeai==0.4.1
python-dotenv==1.0.0